Feat: Modification and Creation of Routes in App.routes
New components were created to show different information in the new routes created in app.routes.

These components will serve as ViewPages to show different information in each View.

We also modified the href of the different buttons in the Sidebar to be able to navigate between the different routes.